Move to custom hook & tests
